backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.30.0
httpx==0.27.0
opencv-python-headless==4.9.0.80
gunicorn==21.2.0